refactor (dashboard): move hook, services and context folders outside app directory
