+++ app/page.tsx
@@ -1,5 +1,5 @@
 import OfficesGrid from "@/components/Grids/OfficesGrid/OfficesGrid";
-import { getOfficesService } from "./services/offices";
+import { getOfficesService } from "../services/offices";
 import OfficeCard from "@/components/cards/OfficeCard/OfficeCard";
 export default async function Home() {
   const offices = await getOfficesService();
